--- test/ast_tests/test.cpp
@@ -1,10 +1,0 @@
-//
-// Created by Yuying on 2022/5/3.
-//
-#include "include/AST.h"
-using namespace std;
-int main()
-{
-    return 0;
-}
-
